Update 0.4 (09.09)

+++ api/src/main/java/org/apache/iceberg/types/Conversions.java
@@ -52,7 +52,7 @@ public class Conversions {
       case LONG:
         return Long.valueOf(asString);
       case FLOAT:
-        return Long.valueOf(asString);
+        return Float.valueOf(asString);
       case DOUBLE:
         return Double.valueOf(asString);
       case DATE:
